refactor: chain mocks now in testing package
