+++ packages/core/package.json
@@ -30,6 +30,7 @@
   "bugs": "https://github.com/KILTprotocol/sdk-js/issues",
   "homepage": "https://github.com/KILTprotocol/sdk-js#readme",
   "devDependencies": {
+    "@kiltprotocol/testing": "workspace:*",
     "@types/uuid": "^8.0.0",
     "rimraf": "^3.0.2",
     "typescript": "^4.5.4"
