+++ packages/core/src/balance/Balance.spec.ts
@@ -13,33 +13,34 @@ import { SubmittableResult } from '@polkadot/api'
 import { GenericAccountIndex as AccountIndex } from '@polkadot/types/generic/AccountIndex'
 import type { AccountData, AccountInfo } from '@polkadot/types/interfaces'
 import { BN } from '@polkadot/util'
-import { BlockchainUtils } from '@kiltprotocol/chain-helpers'
+import {
+  Blockchain,
+  BlockchainApiConnection,
+  BlockchainUtils,
+} from '@kiltprotocol/chain-helpers'
 
 import type { Balances, KeyringPair } from '@kiltprotocol/types'
 import { Keyring } from '@kiltprotocol/utils'
 import { TypeRegistry } from '@polkadot/types'
+import { ApiMocks } from '@kiltprotocol/testing'
 import {
   getBalances,
   listenToBalanceChanges,
   makeTransfer,
 } from './Balance.chain'
 import * as BalanceUtils from './Balance.utils'
-import * as Kilt from '../kilt/Kilt'
 
-jest.mock(
-  '@kiltprotocol/chain-helpers/lib/cjs/blockchainApiConnection/BlockchainApiConnection'
-)
+const mockedApi: any = ApiMocks.getMockedApi()
+const blockchain = new Blockchain(mockedApi)
+BlockchainApiConnection.setConnection(Promise.resolve(blockchain))
 
 const BALANCE = 42
 const FEE = 30
 
 describe('Balance', () => {
-  Kilt.config({ address: 'ws://testSting' })
   const keyring = new Keyring({ type: 'sr25519', ss58Format: 38 })
   let alice: KeyringPair
   let bob: KeyringPair
-  const blockchainApi =
-    require('@kiltprotocol/chain-helpers/lib/cjs/blockchainApiConnection/BlockchainApiConnection').__mocked_api
 
   const accountInfo = (balance: number): AccountInfo => {
     return {
@@ -53,7 +54,7 @@ describe('Balance', () => {
     } as AccountInfo
   }
 
-  blockchainApi.query.system.account = jest.fn(
+  mockedApi.query.system.account = jest.fn(
     (accountAddress, cb): AccountInfo => {
       if (cb) {
         setTimeout(() => {
@@ -105,7 +106,7 @@ describe('Balance', () => {
     const status = await makeTransfer(bob.address, amount, exponent).then(
       (tx) => BlockchainUtils.signAndSubmitTx(tx, alice, { reSign: true })
     )
-    expect(blockchainApi.tx.balances.transfer).toHaveBeenCalledWith(
+    expect(mockedApi.tx.balances.transfer).toHaveBeenCalledWith(
       bob.address,
       expectedAmount
     )
