+++ packages/testing/package.json
@@ -16,7 +16,7 @@
     "lib/**/*"
   ],
   "scripts": {
-    "updateMetadata": "node ./scripts/fetchMetadata.js -o './src/metadata.json' -e 'https://spiritnet.api.onfinality.io/public-http'",
+    "updateMetadata": "node ./scripts/fetchMetadata.js -o './src/mocks/metadata/spiritnet.json' -e 'https://spiritnet.api.onfinality.io/public-http'",
     "clean": "rimraf ./lib",
     "build": "yarn clean && yarn build:ts",
     "build:ts": "yarn build:cjs && yarn build:esm",
@@ -32,6 +32,7 @@
   "bugs": "https://github.com/KILTprotocol/sdk-js/issues",
   "homepage": "https://github.com/KILTprotocol/sdk-js#readme",
   "dependencies": {
+    "@kiltprotocol/types": "workspace:*",
     "@polkadot/api": "^6.11.1",
     "@polkadot/types": "^6.11.1"
   },
