+++ packages/testing/src/mocks/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Copyright 2018-2021 BOTLabs GmbH.
+ *
+ * This source code is licensed under the BSD 4-Clause "Original" license
+ * found in the LICENSE file in the root directory of this source tree.
+ */
+
+export {
+  createAugmentedApi,
+  createRegistryFromMetadata,
+} from './typeRegistry.js'
+export { getMockedApi, MockApiPromise } from './mockedApi.js'
+export { mockChainQueryReturn } from './mockedApi.utils.js'
