+++ packages/testing/src/mocks/metadata/spiritnet.json
